requirements.txt
gradio
groq
PyPDF2
python-docx
sentence-transformers
numpy
scikit-learn